+++ src/App.jsx
@@ -3,9 +3,9 @@ import { Route, Routes } from 'react-router-dom';
 import GlobalStyles from 'styles/globalStyles';
 
 import Layout from 'components/common/Layout';
-import ProductList from 'components/ProductList';
-import ProductDetail from 'components/ProductDetail';
-import NotFound from 'components/NotFound';
+import ProductListPage from 'pages/ProductList';
+import ProductDetailPage from 'pages/ProductDetail';
+import NotFoundPage from 'pages/NotFound';
 
 const App = () => {
   return (
@@ -13,9 +13,9 @@ const App = () => {
       <GlobalStyles />
       <Layout>
         <Routes>
-          <Route path="/react-shopping-cart" element={<ProductList />} />
-          <Route path="/react-shopping-cart/product/:id" element={<ProductDetail />} />
-          <Route path="*" element={<NotFound />} />
+          <Route path="/react-shopping-cart" element={<ProductListPage />} />
+          <Route path="/react-shopping-cart/product/:id" element={<ProductDetailPage />} />
+          <Route path="*" element={<NotFoundPage />} />
         </Routes>
       </Layout>
     </div>
